move merkle helpers to logger package
